Update 'properties' to 'property values'
The task asks to multiply numeric object 'properties', however the term 'object properties' include both the 'property keys' of the object as well as the 'property values' of the object. It is clear the intention of the task is to change property values. 
This change improves clarity of the task description.